Update serialized asset format. Use int16 for asset type id.
Former-commit-id: 7d065adc6ff7194c7a35ddf8d7baa2336659cfbf

+++ plugin/serializer/include/minko/file/AbstractSerializerParser.hpp
@@ -33,9 +33,9 @@ namespace minko
 		{
 		public:
 			typedef std::shared_ptr<AbstractSerializerParser>				Ptr;
-			typedef msgpack::type::tuple<unsigned int, short, std::string>	SerializedAsset;
+			typedef msgpack::type::tuple<short, short, std::string>	SerializedAsset;
 			typedef std::shared_ptr<file::AssetLibrary>						AssetLibraryPtr;
-		
+
 		private:
 			typedef std::shared_ptr<component::JobManager::Job>												JobPtr;
 			typedef std::shared_ptr<Dependency>																DependencyPtr;
